added the undo redo buttons
