+++ app/src/main/java/com/example/mspaint/mainui/drawingCanvas.kt
@@ -1,46 +1,31 @@
 package com.example.mspaint.mainui
 
 import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.gestures.detectDragGestures
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Button
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import com.example.mspaint.R
 import com.example.mspaint.canvasObjectData.Line
+import com.example.mspaint.tools.pencil
 
 @Composable
-fun DrawScreen() {
-    val lines = remember {
-        mutableStateListOf<Line>()
-    }
-    Canvas(
-// sets up the canvas
-        modifier = Modifier
-            .fillMaxSize()
-            .pointerInput(true) {
-                detectDragGestures { change, dragAmount ->
-                    change.consume()
+fun DrawScreen(
+) {
 
-                    val line = Line(
-                        start = change.position - dragAmount,
-                        end = change.position
-                    )
-                    lines.add(line)
-                }
-            }
-    ) // creates the actual drawing. drawLine function is a part of Jetpack Compose
-    {
-        lines.forEach { line ->
-            drawLine(
-                color = line.color,
-                start = line.start,
-                end = line.end,
-                strokeWidth = line.strokeWidth.toPx(),
-                cap = StrokeCap.Round
-            )
-        }
-    }
+
+
 }
+
+
