+++ app/src/main/java/com/example/mspaint/mainui/firstRow.kt
@@ -103,6 +103,7 @@ fun firstRow(): Boolean {
     return hidden
 }
 
+
 @Preview
 @Composable
 fun FirstRowPreview() {
