+++ app/src/main/java/com/example/mspaint/mainui/secondRow.kt
@@ -26,7 +26,10 @@ import com.example.mspaint.tools.pencil
 import com.example.mspaint.ui.theme.Grey
 
 @Composable
-fun SecondRow(){
+fun SecondRow(
+    undo: ()-> Unit,
+    redo: ()-> Unit
+){
     Row(
         modifier = Modifier
             .fillMaxWidth()
@@ -43,7 +46,7 @@ fun SecondRow(){
             modifier = reusableModifier
         ) {
             Button(
-                onClick = { pencil() },
+                onClick = { undo() },
                 shape = RectangleShape,
                 modifier = Modifier.size(width = 80.dp, height = 65.dp)
             ) {
@@ -58,7 +61,7 @@ fun SecondRow(){
             modifier = reusableModifier
         ) {
             Button(
-                onClick = { pencil() },
+                onClick = { redo() },
                 shape = RectangleShape,
                 modifier = Modifier.size(width = 80.dp, height = 65.dp)
             ) {
